python/requirements.txt
strawberry-graphql==0.215.0
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
pytest-asyncio==0.23.3

